fix: error handling for dto mapping for many items
We map to null if decoding fails. This avoids blocking the entire result if some items in the list fail to decode
