+++ apps/web/core/io/dto/triples.ts
@@ -1,0 +1,32 @@
+import { Value } from '~/core/types';
+
+import { SubstreamTriple } from '../schema';
+
+function extractValue(networkTriple: SubstreamTriple): Value {
+  switch (networkTriple.valueType) {
+    case 'TEXT':
+      return { type: 'TEXT', value: networkTriple.textValue };
+    case 'ENTITY': {
+      return {
+        type: 'ENTITY',
+        value: networkTriple.entityValue.id,
+        name: networkTriple.entityValue.name,
+      };
+    }
+    case 'TIME':
+      return { type: 'TIME', value: networkTriple.textValue };
+    case 'URI':
+      return { type: 'URI', value: networkTriple.textValue };
+  }
+}
+
+export function TripleDto(triple: SubstreamTriple) {
+  return {
+    entityId: triple.entity.id,
+    entityName: triple.entity.name,
+    attributeId: triple.attribute.id,
+    attributeName: triple.attribute.name,
+    value: extractValue(triple),
+    space: triple.space.id,
+  };
+}
